+++ composer.json
@@ -91,6 +91,12 @@
             ".\\vendor\\bin\\pint --no-interaction --ansi --dirty"
         ],
         "ci-test": [
+            "@php artisan test --testsuite=CI --compact --parallel --ansi"
+        ],
+        "ci-test:integration": [
+            "@php artisan test --testsuite=Integration --compact --ansi"
+        ],
+        "ci-test:all": [
             "@php artisan test --compact --parallel --ansi"
         ],
         "ci-reset:db": [
