+++ phpunit.xml
@@ -5,16 +5,26 @@
          colors="true"
 >
     <testsuites>
+        <!-- Core API tests (backend only) -->
         <testsuite name="Unit">
             <directory>tests/Unit</directory>
         </testsuite>
         <testsuite name="Feature">
             <directory>tests/Feature</directory>
         </testsuite>
+        <testsuite name="CiCd">
+            <directory>tests/CiCd</directory>
+        </testsuite>
+        
+        <!-- Frontend integration tests (requires npm run build) -->
         <testsuite name="Integration">
             <directory>tests/Integration</directory>
         </testsuite>
-        <testsuite name="CiCd">
+        
+        <!-- Default suite for CI (excludes Integration) -->
+        <testsuite name="CI">
+            <directory>tests/Unit</directory>
+            <directory>tests/Feature</directory>
             <directory>tests/CiCd</directory>
         </testsuite>
     </testsuites>
